monitorapp/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
requests==2.26.0
urllib3==1.26.7